refactor(tags): read in-flight options from a transport-free module
Rows render the overlay on every pass, and importing it from the mutation
module dragged the REST and GraphQL clients — and with them the Soup
websocket — into every consumer that only wanted to display a pending
tag. Under jsdom that surfaced as unhandled "ws does not work in the
browser" rejections from the tag row test.
